Rename interdomain package to clusterinfo
