+++ controlplane/pkg/apis/clusterinfo/gen.go
@@ -1,3 +1,3 @@
-package interdomain
+package clusterinfo
 
 //go:generate bash -c "protoc -I . clusterinfo.proto --go_out=plugins=grpc:. --proto_path=$GOPATH/src/ --proto_path=$GOPATH/pkg/mod/  --proto_path=$( go list -f '{{ .Dir }}' -m github.com/golang/protobuf )"
